Partner-Oculus-4.25 @14415916
OVRPlugin 1.52.1
Platform SDK 1.40.0
Audio SDK 1.41.0
Avatar SDK 1.36.0

+++ Engine/Source/Runtime/Renderer/Private/PostProcess/PostProcessDOF.cpp
@@ -287,7 +287,7 @@ FPooledRenderTargetDesc FRCPassPostProcessDOFSetup::ComputeOutputDesc(EPassOutpu
 	Ret.Extent.Y = FMath::Max(1, Ret.Extent.Y);
 
 	Ret.Reset();
-	Ret.TargetableFlags &= ~(uint32)TexCreate_UAV;
+	Ret.TargetableFlags &= ~TexCreate_UAV;
 	Ret.TargetableFlags |= TexCreate_RenderTargetable;
 	Ret.AutoWritable = false;
 	Ret.DebugName = (InPassOutputId == ePId_Output0) ? TEXT("DOFSetup0") : TEXT("DOFSetup1");
